ADV RW Confirm/requirements.txt
pyinstaller
pycryptodome
pillow
opencv-python
tk
numpy